update init tests to call initializePsychicApp for SSR
unfortunately, nextjs doesn't offer a clean way for us to approach this
problem. We are forced to call initializePsychicApp for each SSR
component accessing a model, which forced my hand at introducing the
`maybeInitializePsychicApp` function, which will only perform
initialization once, and is safe to call repeatedly throughout the app.

This will need to be documented, since it is non-obvious that one would
have to do this.